Extract out common status bar entry management code
